Fixes an issue where trying to hit route endpoints wasn't working.
In docker-chs-dev we have a docker compose which looks for /delta
in a request and routes it towards our container. We forgot to still
add /delta to our route which meant the endpoint wasn't being hit.

By adding /delta we ensure we hit the endpoint and functionality is
now working as expected.

Removed unwanted space between calls in main.