Add ability to publish message onto a Kafka topic
Reworked the OfficerDeltaHandler to now take the REST request
and give it to the new KafkaService for publishing.

Added a common.go helper file which will contain common code shared
between all future handlers. This minimised duplication and makes it
easier to maintain.

Reworked the KafkaService to account for dynamic topic variable and
also improve error handling.

Improved logging across the project (on going process).