Begin unit test coverage implementation for new validation
Rework and include the CH Validator into the Officer Delta Handler
so that we can mock it's calls for unit testing. Also added unit
tests to cover new code in the OfficerDeltaHandler.

Move the CHError model into it's correct model package, and
correctly document it.

Hide CH Validator behind interface so that it can be mocked. Also
created a mock version of the file for unit testing.

Added var to config file to allow unit testing and re-directing of
config validation.

More commits to come for unit testing.