Change routes on validator to use dummy routes
To avoid confusion later in development, we've changed
the routes on the unit tests to dummy routes where possible
and commented why we use real routes in certain areas to avoid
any confusion.

This should help future developers understand that this code
will not need to be duplicated for future deltas.