Use karate style tests to unit test the schema is correctly working.
By creating request bodies and expected response bodies, we have
created unit tests which call the validator with no mocking involved
and asserted that each time we call with a given request, the expected
response is matched directly to what is actually returned.

Implemented custom compare function which compares the actual and
expected responses given to make sure they match or don't match.

Reworked the chOpenApiValidator code to allow for multiErrors to
correctly be handled as the for each statement was returning out of the
method before it could complete it's entire loop.