Rework the handling of returned errors from kin lib
Using the errors.As and errors.Is functions introduced in go
v 1.13 we can now handle errors more gracefully instead of using
a type switch with potential to panic the application.

Added use of these new methods/functions and reworked private
methods in chOpenApiValidator.go to be easier to read and maintain.