Add validate endpoint for officers
Add an endpoint which only validates the request body for officers. This endpoint calls the kin-openapi library to validate and returns a response depending on the outcome of the validation.
Add validate endpoint to register and api-spec
Add new handler that only calls the validation function
Create util test file that has common functionality for handler tests

DSND-96