Add boolean field to officerDeltaHandler to determine whether route is for validation only
To make future deltas to be added easily by only having one handler for all routes, a boolean flag has been added to the officerDeltaHandler to determine if a route is for validation only. If this flag is true then the calls to the functions to send the message to kafka is ignored and only the validation of the data against the OpenApi specification is done.

Delete deltaValidationHandler.go, deltaValidationHandler_test.go and handlersTestUtil.go as no longer needed
Add boolean flag, doValidationOnly, to the struct for officerDeltaHandler and add parameter to constructor to set flag
Add if statement that checks the new flag is false around the calls in the ServeHttp function of officerDeltaHandler.go that send the message to kafka
Update routes in register.go to pass the appropriate flag for validation depending on route

DSND-96