Remove use of sync.Once due to nil pointer issue identified
We identified a nil pointer reference issue which was resulting
in a panic occuring and the application not running correctly.
The issue was due to the way that we created and used the CHValidator service.

When we create the CHValidator service, it contains a nil pointer for it's doc
member variable, this doc pointer is later updated to reference a schema location
in memory when we call to validate a request. We found that using sync.Once loaded the
doc pointer fine for the 1st handler, but when other handlers attempted to load the doc
they were being stopped by sync.Once and then trying to validate using a nil doc pointer.

We have removed the use of sync.Once and instead opted to move the assignment of the doc
to the constructor of the CHValidator, as this constructor is called once in Register.go
on applications start up, it doesn't have to worry about threading issues. We then pass
that fully instantiated object as a copy to each new handler when it is registered
for each endpoint, which means that all endpoints have the document loaded before calling
their validate method.

This caused a lot of unit tests to break as logic had been reworked and reshaped,
So this PR also includes changes to unit tests to fix and update them to correctly
cover the new logic.

Go fmt also ran.