Rework the CH Validator to stop passing duplicate parameters around
It was discovered that we pass the openAPISpec string into the
ValidateAgainstSpec method and also pass it into the New method
for the CH Validator. So we have removed the parameter from the
validate method and inside of the New method, chose to assign it
to a local member variable of the CH Validation struct and call
it directly from the struct when used.

Cleaned up outdated mocks using mockgen

Cleaned up comments and documentation

Go fmt